Remove multiple backends
merge testing into one file
